Initial import into GIT
This is an import of the NetFPGA-1G respository (previously subversion
based). History from subversion is *not* be transferred -- there was too
much cruft in the old tree.

Many people have contributed to this project to make it what it is.
We're sad to lose the old history that showed contributions up to
this point.

Note: This tree will be used for both 1G and 10G code in future.