sim: verify number of reads (bug fix)
An earlier commit verified the number of reads performed during
simulation. The commit didn't take into account the "automatic" reads
that are performed in response to interrupts.

The testbench code automatically reads the interrupt status register
when an interrupt is generated. This may then trigger other reads, such
as reading the DMA ingress status register.